requirements.txt
aiosmtpd==1.4.6
asgiref==3.8.1
atpublic==5.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
crispy-bootstrap5==2024.10
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
idna==3.10
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
urllib3==2.2.3
